idempotent ssh population

+++ cmd/prepare-node/prep-authorized-keys.sh
@@ -8,6 +8,8 @@ set -eux
 srcdir="$(dirname $0)"
 
 mkdir -p $HOME/.ssh
-cp "${srcdir}/barista.pub" $HOME/.ssh/authorized_keys
+if [ ! -f "$HOME/.ssh/authorized_keys" ]; then
+  cp "${srcdir}/barista.pub" $HOME/.ssh/authorized_keys
+fi
 chmod 700 $HOME/.ssh
 chmod 400 $HOME/.ssh/*
